requirements.txt
SQLAlchemy>=1.4
pymssql
python-dotenv
fastapi
uvicorn[standard]
requests
alembic
pytest
httpx